fix: use original filename (not uuid) to set FHIR attachment title
# Ticket 🎫

This closes https://github.com/dot-base/medical-dashboard/issues/1326

# Description 📖
When Formidable [was update from version v1.2.5 to v2.0.5.](https://github.com/dot-base/file-storage-api/pull/46) setting the `title` of FHIR Attachment was refactored. It is set by using the `"newFilename"`. 

This was refactored due to changes that came in with [version 2.0.0](https://github.com/node-formidable/formidable/blob/master/CHANGELOG.md#200): The previously used `file.name` was split into `file.newFilename` and `file.originalFilename`.

But in order to make a color-highlighted visualization of DBS Lead work, we need to keep the `originalFilename` as a attachment title. The attachment title is used in the medical dashboard to distinguish between `"electrode"` or `"anatomy"` ply-files to visualize them color-highlighted

# How to Test? 🧪

Set up dotbase.local and create a new `Neurolokalisation` Document for a patient. Upload the ply files below.
The result should be color-highlighted (not black-and-white).
[Archiv.zip](https://github.com/dot-base/file-storage-api/files/12597330/Archiv.zip)

# Is there something controversial ? 🚨

We need to write a migration (see [fhir-resources/#449](https://github.com/dot-base/fhir-resources/issues/449)) to make this work for already uploaded Lead DBS files. 

# References (optional) 🔗
none

# Checklist ✅
- [x] My changes generate no new warnings
- [x] Made corresponding changes to the documentation if applicable
- [x] Commented my code, particularly in hard-to-understand areas if applicable
- [x] Performed a self-review of my own changes
- [ ] Added relevant unit tests